+++ featuretools/computational_backends/feature_tree.py
@@ -8,7 +8,7 @@ from ..utils import gen_utils as utils
 
 from featuretools import variable_types
 from featuretools.exceptions import UnknownFeature
-from featuretools.primitives.primitive_utils import (
+from featuretools.primitives.base import (
     AggregationPrimitive,
     DirectFeature,
     IdentityFeature,
