+++ featuretools/primitives/api.py
@@ -1,11 +1,13 @@
 # flake8: noqa
-from .builtin_primitives.aggregation_primitives import *
-from .builtin_primitives.binary_transform import *
-from .builtin_primitives.cum_transform_feature import *
-from .builtin_primitives.transform_primitive import *
-from .builtin_primitives.utils import (
+from .base import Feature
+from .install import install_primitives
+from .installed import *
+from .standard.aggregation_primitives import *
+from .standard.binary_transform import *
+from .standard.cum_transform_feature import *
+from .standard.transform_primitive import *
+from .utils import (
     get_aggregation_primitives,
     get_transform_primitives,
     list_primitives
 )
-from .installed import *
