+++ featuretools/primitives/base/api.py
@@ -3,12 +3,6 @@ from .aggregation_primitive_base import (
     AggregationPrimitive,
     make_agg_primitive
 )
-from .install import (
-    get_installation_dir,
-    install_primitives,
-    list_primitive_files,
-    load_primitive_from_file
-)
 from .primitive_base import (
     DirectFeature,
     Feature,
