+++ featuretools/primitives/install.py
@@ -8,7 +8,7 @@ from inspect import isclass
 from smart_open import smart_open
 from tqdm import tqdm
 
-from .primitive_base import PrimitiveBase
+from .base.primitive_base import PrimitiveBase
 
 import featuretools
