+++ featuretools/primitives/standard/binary_transform.py
@@ -3,11 +3,10 @@ from builtins import str
 
 import numpy as np
 
-from ..primitive_utils.primitive_base import PrimitiveBase
-from ..primitive_utils.transform_primitive_base import (
-    TransformPrimitive
-)
-from ..primitive_utils.utils import apply_dual_op_from_feat
+from ..base.primitive_base import PrimitiveBase
+from ..base.transform_primitive_base import TransformPrimitive
+from .utils import apply_dual_op_from_feat
+
 from featuretools.variable_types import (
     Boolean,
     Datetime,
