+++ featuretools/primitives/standard/transform_primitive.py
@@ -6,10 +6,11 @@ from builtins import str
 import numpy as np
 import pandas as pd
 
-from ..primitive_utils.transform_primitive_base import (
+from ..base.transform_primitive_base import (
     TransformPrimitive,
     make_trans_primitive
 )
+
 from featuretools.variable_types import (
     Boolean,
     Datetime,
