+++ featuretools/tests/computational_backend/test_calculate_feature_matrix.py
@@ -26,12 +26,12 @@ from featuretools.computational_backends.utils import (
     get_next_chunk,
     n_jobs_to_workers
 )
-from featuretools.primitive_utils import (
+from featuretools.primitives import Count, Max, Min, Percentile, Sum
+from featuretools.primitives.base import (
     AggregationPrimitive,
     DirectFeature,
     IdentityFeature
 )
-from featuretools.primitives import Count, Max, Min, Percentile, Sum
 
 
 @pytest.fixture(scope='module')
