+++ featuretools/tests/computational_backend/test_pandas_backend.py
@@ -12,7 +12,6 @@ from ..testing_utils import make_ecommerce_entityset
 import featuretools as ft
 from featuretools import Timedelta
 from featuretools.computational_backends.pandas_backend import PandasBackend
-from featuretools.primitive_utils import DirectFeature
 from featuretools.primitives import (
     And,
     Count,
@@ -30,6 +29,7 @@ from featuretools.primitives import (
     Sum,
     Trend
 )
+from featuretools.primitives.base import DirectFeature
 
 
 @pytest.fixture(scope='module')
