+++ featuretools/tests/dfs_tests/test_deep_feature_synthesis.py
@@ -11,14 +11,6 @@ from pympler.asizeof import asizeof
 from ..testing_utils import feature_with_name, make_ecommerce_entityset
 
 import featuretools as ft
-from featuretools.primitive_utils import (
-    AggregationPrimitive,
-    DirectFeature,
-    Feature,
-    IdentityFeature,
-    TransformPrimitive,
-    make_agg_primitive
-)
 from featuretools.primitives import (
     Absolute,
     Add,
@@ -31,6 +23,14 @@ from featuretools.primitives import (
     Mode,
     Sum,
     TimeSincePrevious
+)
+from featuretools.primitives.base import (
+    AggregationPrimitive,
+    DirectFeature,
+    Feature,
+    IdentityFeature,
+    TransformPrimitive,
+    make_agg_primitive
 )
 from featuretools.synthesis import DeepFeatureSynthesis
 from featuretools.utils.pickle_utils import save_obj_pickle
