+++ featuretools/tests/primitive_tests/test_agg_feats.py
@@ -9,11 +9,6 @@ import pytest
 from ..testing_utils import feature_with_name, make_ecommerce_entityset
 
 import featuretools as ft
-from featuretools.primitive_utils import (
-    AggregationPrimitive,
-    Feature,
-    make_agg_primitive
-)
 from featuretools.primitives import (
     Count,
     Mean,
@@ -23,6 +18,11 @@ from featuretools.primitives import (
     Sum,
     TimeSinceLast,
     get_aggregation_primitives
+)
+from featuretools.primitives.base import (
+    AggregationPrimitive,
+    Feature,
+    make_agg_primitive
 )
 from featuretools.synthesis.deep_feature_synthesis import (
     DeepFeatureSynthesis,
