+++ featuretools/tests/primitive_tests/test_direct_features.py
@@ -3,7 +3,7 @@ import pytest
 from ..testing_utils import make_ecommerce_entityset
 
 from featuretools.computational_backends import PandasBackend
-from featuretools.primitive_utils import DirectFeature
+from featuretools.primitives.base import DirectFeature
 
 
 @pytest.fixture(scope='module')
