+++ featuretools/tests/primitive_tests/test_primitive_base.py
@@ -3,8 +3,8 @@ from pympler.asizeof import asizeof
 
 from ..testing_utils import make_ecommerce_entityset
 
-from featuretools.primitive_utils import Feature, IdentityFeature
 from featuretools.primitives import Last, Mode, Sum
+from featuretools.primitives.base import Feature, IdentityFeature
 from featuretools.variable_types import Datetime, Numeric
